Refactor interface code a little; tests passing
