+++ server/game/Interfaces.ts
@@ -6,31 +6,25 @@ import type CardAbility from './core/ability/CardAbility';
 import type { IAttackProperties } from './gameSystems/AttackSystem';
 import type { RelativePlayer, TargetMode, CardType, Location, EventName, PhaseName, LocationFilter } from './core/Constants';
 import type { GameEvent } from './core/event/GameEvent';
-// import type { StatusToken } from './StatusToken';
+import type { IActionTarget, IActionTargets, ITriggeredAbilityTarget, ITriggeredAbilityTargets } from './TargetInterfaces';
 
-interface IBaseTarget {
-    activePromptTitle?: string;
+// ********************************************** EXPORTED TYPES **********************************************
+export type ITriggeredAbilityProps = ITriggeredAbilityWhenProps | ITriggeredAbilityAggregateWhenProps;
+
+export interface IConstantAbilityProps<Source = any> {
+    title: string;
     locationFilter?: LocationFilter | LocationFilter[];
-    controller?: ((context: AbilityContext) => RelativePlayer) | RelativePlayer;
-    player?: ((context: AbilityContext) => RelativePlayer) | RelativePlayer;
-    hideIfNoLegalTargets?: boolean;
-    gameSystem?: GameSystem | GameSystem[];
-}
+    // TODO: what's the difference between condition and match?
+    condition?: (context: AbilityContext<Source>) => boolean;
+    match?: (card: Card, context?: AbilityContext<Source>) => boolean;
+    targetController?: RelativePlayer;
+    targetLocation?: Location;
 
-type IChoicesInterface = Record<string, ((context: AbilityContext) => boolean) | GameSystem | GameSystem[]>;
+    // TODO: can we get a real signature here
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-function-type
+    effect: Function | Function[];
 
-interface ITargetSelect extends IBaseTarget {
-    mode: TargetMode.Select;
-    choices: (IChoicesInterface | object) | ((context: AbilityContext) => IChoicesInterface | object);
-    condition?: (context: AbilityContext) => boolean;
-    targets?: boolean;
-}
-
-interface ITargetAbility extends IBaseTarget {
-    mode: TargetMode.Ability;
-    cardType?: CardType | CardType[];
-    cardCondition?: (card: Card, context?: AbilityContext) => boolean;
-    abilityCondition?: (ability: CardAbility) => boolean;
+    createCopies?: boolean;
 }
 
 export interface IInitiateAttack extends IAttackProperties {
@@ -40,71 +34,50 @@ export interface IInitiateAttack extends IAttackProperties {
     targetCondition?: (card: Card, context: TriggeredAbilityContext) => boolean;
 }
 
-// interface TargetToken extends BaseTarget {
-//     mode: TargetMode.Token;
-//     optional?: boolean;
-//     locationFilter?: Location | Location[];
-//     cardType?: CardType | CardType[];
-//     singleToken?: boolean;
-//     cardCondition?: (card: BaseCard, context?: AbilityContext) => boolean;
-//     tokenCondition?: (token: StatusToken, context?: AbilityContext) => boolean;
-// }
+export interface IActionProps<Source = any> extends IAbilityProps<AbilityContext<Source>> {
+        condition?: (context?: AbilityContext<Source>) => boolean;
 
-interface IBaseTargetCard extends IBaseTarget {
-    cardType?: CardType | CardType[];
-    locationFilter?: Location | Location[];
-    optional?: boolean;
-}
+        /**
+         * If true, any player can trigger the ability. If false, only the card's controller can trigger it.
+         */
+        anyPlayer?: boolean;
+        phase?: PhaseName | 'any';
+    }
 
-interface ITargetCardExactlyUpTo extends IBaseTargetCard {
-    mode: TargetMode.Exactly | TargetMode.UpTo;
-    numCards: number;
-    sameDiscardPile?: boolean;
-}
+export type traitLimit = Record<string, number>;
 
-interface ITargetCardExactlyUpToVariable extends IBaseTargetCard {
-    mode: TargetMode.ExactlyVariable | TargetMode.UpToVariable;
-    numCardsFunc: (context: AbilityContext) => number;
-}
-
-interface ITargetCardMaxStat extends IBaseTargetCard {
-    mode: TargetMode.MaxStat;
-    numCards: number;
-    cardStat: (card: Card) => number;
-    maxStat: () => number;
-}
-
-interface TargetCardSingleUnlimited extends IBaseTargetCard {
-    mode?: TargetMode.Single | TargetMode.Unlimited;
-}
-
-type ITargetCard =
-    | ITargetCardExactlyUpTo
-    | ITargetCardExactlyUpToVariable
-    | ITargetCardMaxStat
-    | TargetCardSingleUnlimited
-    | ITargetAbility;
-    // | TargetToken;
-
-interface ISubTarget {
-    dependsOn?: string;
-}
-
-interface IActionCardTarget {
-    cardCondition?: (card: Card, context?: AbilityContext) => boolean;
-}
-
-type IActionTarget = (ITargetCard & IActionCardTarget) | ITargetSelect | ITargetAbility;
-
-type IActionTargets = Record<string, IActionTarget & ISubTarget>;
-
-type EffectArg =
+export type EffectArg =
     | number
     | string
     | RelativePlayer
     | Card
     | { id: string; label: string; name: string; facedown: boolean; type: CardType }
     | EffectArg[];
+
+export type WhenType = {
+        [EventNameValue in EventName]?: (event: any, context?: TriggeredAbilityContext) => boolean;
+    };
+
+// TODO UPGRADES
+// export interface IAttachmentConditionProps {
+//     limit?: number;
+//     myControl?: boolean;
+//     opponentControlOnly?: boolean;
+//     unique?: boolean;
+//     faction?: string | string[];
+//     trait?: string | string[];
+//     limitTrait?: traitLimit | traitLimit[];
+//     cardCondition?: (card: Card) => boolean;
+// }
+
+// ********************************************** INTERNAL TYPES **********************************************
+interface ITriggeredAbilityWhenProps extends ITriggeredAbilityBaseProps {
+    when: WhenType;
+}
+
+interface ITriggeredAbilityAggregateWhenProps extends ITriggeredAbilityBaseProps {
+    aggregateWhen: (events: GameEvent[], context: TriggeredAbilityContext) => boolean;
+}
 
 interface IAbilityProps<Context> {
     title: string;
@@ -131,33 +104,9 @@ interface IAbilityProps<Context> {
     then?: ((context?: AbilityContext) => object) | object;
 }
 
-export interface IActionProps<Source = any> extends IAbilityProps<AbilityContext<Source>> {
-    condition?: (context?: AbilityContext<Source>) => boolean;
-
-    /**
-     * If true, any player can trigger the ability. If false, only the card's controller can trigger it.
-     */
-    anyPlayer?: boolean;
-    phase?: PhaseName | 'any';
-}
-
-interface ITriggeredAbilityCardTarget {
-    cardCondition?: (card: Card, context?: TriggeredAbilityContext) => boolean;
-}
-
-type TriggeredAbilityTarget =
-    | (ITargetCard & ITriggeredAbilityCardTarget)
-    | ITargetSelect;
-
-type ITriggeredAbilityTargets = Record<string, TriggeredAbilityTarget & ISubTarget & TriggeredAbilityTarget>;
-
-export type WhenType = {
-    [EventNameValue in EventName]?: (event: any, context?: TriggeredAbilityContext) => boolean;
-};
-
 interface ITriggeredAbilityBaseProps extends IAbilityProps<TriggeredAbilityContext> {
     collectiveTrigger?: boolean;
-    target?: TriggeredAbilityTarget & TriggeredAbilityTarget;
+    target?: ITriggeredAbilityTarget;
     targets?: ITriggeredAbilityTargets;
     handler?: (context: TriggeredAbilityContext) => void;
     then?: ((context?: TriggeredAbilityContext) => object) | object;
@@ -173,45 +122,6 @@ interface ITriggeredAbilityBaseProps extends IAbilityProps<TriggeredAbilityConte
      * 'Pass' button on resolution) or if it is mandatory
      */
     optional?: boolean;
-}
-
-export interface ITriggeredAbilityWhenProps extends ITriggeredAbilityBaseProps {
-    when: WhenType;
-}
-
-export interface ITriggeredAbilityAggregateWhenProps extends ITriggeredAbilityBaseProps {
-    aggregateWhen: (events: GameEvent[], context: TriggeredAbilityContext) => boolean;
-}
-
-export type ITriggeredAbilityProps = ITriggeredAbilityWhenProps | ITriggeredAbilityAggregateWhenProps;
-
-export interface IConstantAbilityProps<Source = any> {
-    title: string;
-    locationFilter?: LocationFilter | LocationFilter[];
-    // TODO: what's the difference between condition and match?
-    condition?: (context: AbilityContext<Source>) => boolean;
-    match?: (card: Card, context?: AbilityContext<Source>) => boolean;
-    targetController?: RelativePlayer;
-    targetLocation?: Location;
-
-    // TODO: can we get a real signature here
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-function-type
-    effect: Function | Function[];
-
-    createCopies?: boolean;
-}
-
-export type traitLimit = Record<string, number>;
-
-export interface IAttachmentConditionProps {
-    limit?: number;
-    myControl?: boolean;
-    opponentControlOnly?: boolean;
-    unique?: boolean;
-    faction?: string | string[];
-    trait?: string | string[];
-    limitTrait?: traitLimit | traitLimit[];
-    cardCondition?: (card: Card) => boolean;
 }
 
 // export type Token = HonoredToken | DishonoredToken;
